requirements.txt
# Core dependencies
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
datetime

# Optional dependencies - commented out to avoid compatibility issues
# sentence-transformers==2.2.2
# scikit-learn==1.2.2

# For development
pytest==7.4.0
black==23.7.0